+++ application/src/main/java/com/ericsson/bss/cassandra/ecchronos/application/DefaultNativeConnectionProvider.java
@@ -41,7 +41,6 @@ public class DefaultNativeConnectionProvider implements NativeConnectionProvider
     private static final Logger LOG = LoggerFactory.getLogger(DefaultNativeConnectionProvider.class);
 
     private final LocalNativeConnectionProvider myLocalNativeConnectionProvider;
-
     public DefaultNativeConnectionProvider(final Config config,
                                            final Supplier<Security.CqlSecurity> cqlSecuritySupplier,
                                            final CertificateHandler certificateHandler,
@@ -49,6 +48,7 @@ public class DefaultNativeConnectionProvider implements NativeConnectionProvider
                                            final MeterRegistry meterRegistry)
     {
         NativeConnection nativeConfig = config.getConnectionConfig().getCqlConnection();
+        String repairHistoryKeyspace = config.getRepairConfig().getRepairHistory().getKeyspaceName();
         String host = nativeConfig.getHost();
         int port = nativeConfig.getPort();
         boolean remoteRouting = nativeConfig.getRemoteRouting();
@@ -81,7 +81,8 @@ public class DefaultNativeConnectionProvider implements NativeConnectionProvider
                 .withMetricsEnabled(config.getStatisticsConfig().isEnabled())
                 .withMeterRegistry(meterRegistry)
                 .withSchemaChangeListener(defaultRepairConfigurationProvider)
-                .withNodeStateListener(defaultRepairConfigurationProvider);
+                .withNodeStateListener(defaultRepairConfigurationProvider)
+                .withRepairHistoryKeyspace(repairHistoryKeyspace);
 
         myLocalNativeConnectionProvider = establishConnection(nativeConnectionBuilder,
                 host, port, nativeConfig.getTimeout().getConnectionTimeout(TimeUnit.MILLISECONDS),
@@ -194,4 +195,5 @@ public class DefaultNativeConnectionProvider implements NativeConnectionProvider
     {
         myLocalNativeConnectionProvider.close();
     }
+
 }