+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Version 6.0.8
 
 * Refactor ecchronos-binary Test Infrastructure - Issue #1093
+* Schedule not showing if ecchronos keyspace is not replicated correctly - Issue #1112
 
 ## Version 6.0.7